requirements.txt
pandas
openpyxl
Pillow
gTTS
google-cloud-texttospeech
moviepy
customtkinter
langdetect
